version1/requirements.txt
streamlit
azure-cognitiveservices-speech
openai
python-dotenv